Extract core code from WordList manager
