+++ fontforge/wordlistparser.h
@@ -25,31 +25,14 @@
 *******************************************************************************
 *******************************************************************************
 ******************************************************************************/
+#pragma once
 
-#ifndef FONTFORGE_WORDLISTPARSER_H
-#define FONTFORGE_WORDLISTPARSER_H
-
-#include "ffglib.h"
 #include "splinefont.h"
 
 const char* Wordlist_getSCName( SplineChar* sc );
 
 typedef int (*WordlistEscapedInputStringToRealString_getFakeUnicodeOfScFunc)( SplineChar *sc, void* udata );
 extern int WordlistEscapedInputStringToRealString_getFakeUnicodeAsScUnicodeEnc( SplineChar *sc, void* udata );
-
-
-extern unichar_t* WordlistEscapedInputStringToRealStringBasic(
-    SplineFont* sf,
-    unichar_t* input_const,
-    GArray** selected_out );
-
-extern GTextInfo** WordlistLoadFileToGTextInfo( int type, int words_max );
-extern GTextInfo** WordlistLoadFileToGTextInfoBasic( int words_max );
-
-extern void Wordlist_MoveByOffset( GGadget* g, int* idx, int offset );
-extern void Wordlist_touch( GGadget* g );
-
-extern void WordlistLoadToGTextInfo( GGadget* g, int* idx  );
 
 extern void WordlistTrimTrailingSingleSlash( unichar_t* txt );
 
@@ -123,5 +106,3 @@ extern WordListLine WordlistEscapedInputStringToParsedDataComplex(
 /* The unput may contain unencoded glyphs references as \<glyph_name>. These
  * glyphs are recoded to fake values above the Unicode range. */
 extern char* WordlistEscapedInputStringToUTF8(SplineFont* sf, const char* input);
-
-#endif /* FONTFORGE_WORDLISTPARSER_H */
