+++ fontforgeexe/wordlistparser_ui.h
@@ -1,0 +1,39 @@
+/******************************************************************************
+*******************************************************************************
+*******************************************************************************
+
+    Copyright (C) 2013 Ben Martin
+    
+    This file is part of FontForge.
+
+    FontForge is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation,either version 3 of the License, or
+    (at your option) any later version.
+
+    FontForge is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with FontForge.  If not, see <http://www.gnu.org/licenses/>.
+
+    For more details see the COPYING.gplv3 file in the root directory of this
+    distribution.
+
+*******************************************************************************
+*******************************************************************************
+******************************************************************************/
+#pragma once
+
+#include "wordlistparser.h"
+#include "ggadget.h"
+
+extern GTextInfo** WordlistLoadFileToGTextInfo( int type, int words_max );
+extern GTextInfo** WordlistLoadFileToGTextInfoBasic( int words_max );
+
+extern void Wordlist_MoveByOffset( GGadget* g, int* idx, int offset );
+extern void Wordlist_touch( GGadget* g );
+
+extern void WordlistLoadToGTextInfo( GGadget* g, int* idx  );
